Combined System/Indicator-based System/requirements.txt
openai==1.68.2
pandas==2.2.3
yfinance==0.2.54
colorama==0.4.6
python-dotenv==1.0.1
